Database, rushhour and intervals changes
-Start date and end date of an interval now get saved in the database.
-Rushhours can now be edited and deleted
-Intervals can now be deleted
-When updating the timetable, intervals weren't being recalculated
correctly. Right now the intervals are implemented as a sort of
constant, a deleted trip in an interval will just get readded on the
next update. This can be changed so the updater only deletes old trips
and adds new trips (only in the timespan between last system update and now)
without touching trips that have already been added and may have been
changed.